app/pyproject.toml
[project]
name = "steampredictor-app"
version = "1.0.0"
description = "SteamPredictor - Web application for Steam game predictions"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "jinja2>=3.1.6",
    "pydantic>=2.10.6",
    "joblib>=1.5.0",
    "pandas>=3.0.2",
    "pyarrow>=24.0.0",
    "scikit-learn>=1.8.0",
    "torch>=2.11.0",
    "requests>=2.33.1",
    "pillow>=12.2.0",
    "openai-clip>=1.0.1",
    "sentence-transformers>=5.4.1",
    "google-api-python-client>=2.194.0",
    "python-dotenv>=1.2.2",
]
